pyproject.toml
[project]
name = "correct_language"
version = "0.1.0"
description = "Validates that an LLM-generated text is in the expected language."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "fast-langdetect",
    "guardrails-ai>=0.3.2",
    "iso-language-codes",
    "torch",
    "transformers",
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]